# backend/Document/Swagger.yaml
openapi: 3.0.4
info:
  title: National Football League Management API
  description: |-
    This is an API for managing a national football league, built on the OpenAPI 3.0 specification. It provides endpoints for managing team information, players, match schedules, results, and league standings.  
    Learn more about OpenAPI at [https://swagger.io](https://swagger.io). We follow a design-first approach, allowing you to contribute to improving the API, from its definition to its code.  
    Some useful links:  
    - [National Football League API Repository](https://github.com/truonghoangkhiem/QL_Giai_BD_QG_SE104_UIT.git)
  termsOfService: https://swagger.io/terms/
  contact:
    email: truongkhiemhoang@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
...
servers:
  - url: http://localhost:5000
...
paths:
...
  /api/player_results:
    post:
      tags:
        - playerresults
      summary: Create a new player result
      description: Creates a new player result for a specific player, season, and team
      operationId: createPlayerResults
      security:
        - bearerAuth: []
      requestBody:
        description: Information needed to create a player result
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerResultRequest"
      responses:
        "201":
          description: Player result created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid input or player result already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Player or season not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/player_results/season/{seasonid}:
    get:
      tags:
        - playerresults
      summary: Get player results by season ID and date
      description: Retrieves the latest player results for a specific season up to a given date
      operationId: getPlayerResultbySeasonIdAndDate
      parameters:
        - $ref: "#/components/parameters/SeasonId"
      requestBody:
        description: Date to filter player results
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilterDateRequest"
      responses:
        "200":
          description: Latest player results retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerResultsResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No player results found for this season and date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/player_results/player/{playerid}:
    get:
      tags:
        - playerresults
      summary: Get player result by player ID
      description: Retrieves a player result by player ID
      operationId: getPlayerResultsById
      parameters:
        - $ref: "#/components/parameters/PlayerId"
      responses:
        "200":
          description: Player result found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerResultResponse"
        "400":
          description: Invalid player ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Player result not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/player_results/match/{matchid}:
    put:
      tags:
        - playerresults
      summary: Update player results after a match
      description: Updates player results for all players in a match based on match details
      operationId: updatePlayerResultsafterMatch
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MatchId"
      responses:
        "200":
          description: Player results updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid match ID or match season_id is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Match not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/player_results/{id}:
    put:
      tags:
        - playerresults
      summary: Update a player result manually
      description: Updates an existing player result with provided details
      operationId: updatePlayerResults
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlayerResultId"
      requestBody:
        description: Information needed to update the player result
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerResultRequest"
      responses:
        "200":
          description: Player result updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Player result not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - playerresults
      summary: Delete a player result
      description: Deletes an existing player result by its ID
      operationId: deletePlayerResults
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlayerResultId"
      responses:
        "200":
          description: Player result deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid player result ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Player result not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SeasonId:
      name: season_id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d2117c"
...
    MatchId:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d2117b"
...
    PlayerId:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d2117f"
...
    PlayerResultId:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d21180"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Invalid email address"
      required:
        - status
        - message
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
          nullable: true
          example: null
      required:
        - status
        - message
...
    PlayerResult:
      type: object
      properties:
        _id:
          type: string
          example: "67d0ec1a412c1edc00d21180"
        player_id:
          type: string
          example: "67d0ec1a412c1edc00d2117f"
        season_id:
          type: string
          example: "67d0ec1a412c1edc00d2117c"
        team_id:
          type: string
          example: "67d0ec1a412c1edc00d2117d"
        matchesPlayed:
          type: integer
          example: 5
        goals:
          type: integer
          example: 3
        assists:
          type: integer
          example: 2
        yellowCards:
          type: integer
          example: 1
        redCards:
          type: integer
          example: 0
        date:
          type: string
          format: date-time
          example: "2025-04-08T00:00:00Z"
        createdAt:
          type: string
          format: date-time
          example: "2025-04-08T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-04-08T12:00:00Z"
      required:
        - _id
        - player_id
        - season_id
        - team_id
        - matchesPlayed
        - goals
        - assists
        - yellowCards
        - redCards
        - date
        - createdAt
        - updatedAt
    PlayerResultsResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Fetched player results successfully"
        data:
          type: array
          items:
            $ref: "#/components/schemas/PlayerResult"
      required:
        - status
        - message
        - data
    PlayerResultResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Player result found successfully"
        data:
          $ref: "#/components/schemas/PlayerResult"
      required:
        - status
        - message
        - data
    CreatePlayerResultRequest:
      type: object
      properties:
        player_id:
          type: string
          example: "67d0ec1a412c1edc00d2117f"
        season_id:
          type: string
          example: "67d0ec1a412c1edc00d2117c"
        team_id:
          type: string
          example: "67d0ec1a412c1edc00d2117d"
      required:
        - player_id
        - season_id
        - team_id
    UpdatePlayerResultRequest:
      type: object
      properties:
        matchesPlayed:
          type: integer
          example: 6
        goals:
          type: integer
          example: 4
        assists:
          type: integer
          example: 3
        yellowCards:
          type: integer
          example: 2
        redCards:
          type: integer
          example: 0
        date:
          type: string
          format: date-time
          example: "2025-04-08T00:00:00Z"
...
    FilterDateRequest:
      type: object
      properties:
        date:
          type: string
          format: date-time
          example: "2025-04-08T00:00:00Z"
      required:
        - date